Abstract audio pipelines of wav folders
